test: fix Java 11 test compatibility and CI failures
Enable required annotation processors, silence Java 11 compiler warnings, and improve CCM cleanup logging so integration failures are easier to diagnose.

Update tests for newer Mockito behavior by avoiding shared MockitoJUnitRunner usage in parallel integration tests and by verifying cache interactions explicitly where the mock behavior changed.

Fix CI failures exposed by the updated validation matrix: provide Micrometer histogram bounds in the no-SLA test, register config-change listeners before manual reloads, avoid assuming a transient one-connection node state before reconnection wins the race, and validate prepared-statement invalidation through observable cache behavior instead of a test-only removal-listener latch.

+++ metrics/micrometer/src/main/java/com/datastax/oss/driver/internal/metrics/micrometer/MicrometerMetricUpdater.java
@@ -55,14 +55,14 @@ public abstract class MicrometerMetricUpdater<MetricT> extends AbstractMetricUpd
   @Override
   public void incrementCounter(MetricT metric, @Nullable String profileName, long amount) {
     if (isEnabled(metric, profileName)) {
-      getOrCreateCounterFor(metric).increment(amount);
+      getOrCreateCounterFor(metric).increment((double) amount);
     }
   }
 
   @Override
   public void updateHistogram(MetricT metric, @Nullable String profileName, long value) {
     if (isEnabled(metric, profileName)) {
-      getOrCreateDistributionSummaryFor(metric).record(value);
+      getOrCreateDistributionSummaryFor(metric).record((double) value);
     }
   }
 
@@ -70,7 +70,7 @@ public abstract class MicrometerMetricUpdater<MetricT> extends AbstractMetricUpd
   public void markMeter(MetricT metric, @Nullable String profileName, long amount) {
     if (isEnabled(metric, profileName)) {
       // There is no meter type in Micrometer, so use a counter
-      getOrCreateCounterFor(metric).increment(amount);
+      getOrCreateCounterFor(metric).increment((double) amount);
     }
   }
